requirements.txt
Django>=5.0
djangorestframework
django-filter
djangorestframework-simplejwt
drf-yasg
psycopg2-binary
